tickets list small component refactor with types
