+++ src/components/ticket-list/Tickets.tsx
@@ -1,0 +1,150 @@
+import { Center, Spinner } from "@chakra-ui/react"
+import axios from "axios"
+import { useRouter } from "next/router"
+import React, { useEffect, useState } from "react"
+import { useQuery, useQueryClient } from "react-query"
+import {
+  LocationTag,
+  NeedTagType,
+  TICKET_STATUS,
+} from "../../services/ticket.type"
+import { getRootContainer } from "../../services/_root-container"
+import { TicketsListMetaData } from "./TicketsListMetaData"
+import { Pagination } from "./Pagination"
+import { TicketsListHeader } from "./TicketsListHeader"
+import { TicketsListFilters } from "./TicketsListFilters"
+import { TicketsList } from "./TicketsList"
+const ts = getRootContainer().containers.ticketService
+
+export const TRANSPORT_TAG = 5
+
+export const Tickets = ({
+  mineOnly,
+  ticketStatus,
+  title,
+  tags,
+  locationTags,
+  currentTag,
+}: {
+  mineOnly?: boolean
+  ticketStatus: TICKET_STATUS
+  title: string
+  tags: NeedTagType[]
+  locationTags: LocationTag[]
+  currentTag?: NeedTagType
+}) => {
+  const router = useRouter()
+
+  const [selectedTag, setSelectedTag] = useState(
+    parseInt((router.query.tag as string) || "0")
+  )
+  const [selectedPage, setSelectedPage] = useState(
+    parseInt(router.query.page as string) || 1
+  )
+  const [whereToTag, setWhereToTag] = useState(
+    parseInt((router.query["where_to"] as string) || "0")
+  )
+  const [whereFromTag, setWhereFromTag] = useState(
+    parseInt((router.query["where_from"] as string) || "0")
+  )
+  const [queryKey] = useState("tickets")
+
+  const {
+    data: ticketsData,
+    isLoading,
+    isRefetching,
+    refetch,
+  } = useQuery(
+    queryKey,
+    () => {
+      const url = `/api/get-tickets`
+
+      return axios
+        .get(url, {
+          params: {
+            mineOnly: mineOnly,
+            ticketStatus: ticketStatus,
+            tagId: selectedTag,
+            page: selectedPage,
+            whereFromTag: whereFromTag,
+            whereToTag: whereToTag,
+          },
+        })
+        .then((response) => {
+          return response.data
+        })
+    },
+    {
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+      refetchOnMount: false,
+    }
+  )
+
+  const queryClient = useQueryClient()
+
+  useEffect(() => {
+    const tag = parseInt((router.query.tag as string) || "0")
+    if (
+      tag !== TRANSPORT_TAG &&
+      (router.query.where_to || router.query.where_from)
+    ) {
+      delete router.query.where_to
+      delete router.query.where_from
+      router.push(router)
+    }
+    setSelectedTag(parseInt((router.query.tag as string) || "0"))
+  }, [router.query.tag])
+
+  useEffect(() => {
+    setSelectedPage(parseInt(router.query.page as string) || 1)
+  }, [router.query.page])
+
+  useEffect(() => {
+    setWhereToTag(parseInt((router.query["where_to"] as string) || "0"))
+  }, [router.query["where_to"]])
+
+  useEffect(() => {
+    setWhereFromTag(parseInt((router.query["where_from"] as string) || "0"))
+  }, [router.query["where_from"]])
+
+  useEffect(() => {
+    queryClient.cancelQueries(queryKey)
+    void refetch()
+  }, [
+    selectedTag,
+    selectedPage,
+    whereToTag,
+    whereFromTag,
+    queryKey,
+    refetch,
+    queryClient,
+  ])
+
+  return (
+    <>
+      <TicketsListMetaData tag={currentTag} ticketsListTitle={title} />
+      <TicketsListHeader title={title} count={ticketsData?.meta.filter_count} />
+
+      <div className="py-2 mx-auto max-w-7xl sm:px-6 xl:px-0">
+        <TicketsListFilters
+          tags={tags}
+          locationTags={locationTags}
+          selectedTag={selectedTag}
+          whereFromTag={whereFromTag}
+          whereToTag={whereToTag}
+        />
+
+        {isLoading || isRefetching ? (
+          <Center h="100px" color="yellow.400">
+            <Spinner size="xl" thickness="6px" />
+          </Center>
+        ) : (
+          <TicketsList tickets={ticketsData.tickets} />
+        )}
+
+        <Pagination ticketsData={ticketsData} selectedPage={selectedPage} />
+      </div>
+    </>
+  )
+}
