+++ src/services/ticket.type.ts
@@ -1,3 +1,5 @@
+import { Nullable } from "tsdef"
+
 export type Organization = {
   id: number
   name: string
@@ -13,12 +15,12 @@ export interface NeedTagType {
   sort: number
 
   // Human names
-  translation_uk_UA: string
-  translation_en_US: null
-  // user_created: string
-  // date_created: string
-  // user_updated: string
-  // date_updated: string
+  translation_uk_UA: Nullable<string>
+  translation_en_US: Nullable<string>
+  user_created: string
+  date_created: string
+  user_updated: string
+  date_updated: string
 }
 
 export interface LocationTag {
